// src/api/auth/tags.ts
export enum Tags {
  POST_REQUEST_REGISTER = 'POST_REQUEST_REGISTER',
  POST_CONFIRM_REGISTER = 'POST_CONFIRM_REGISTER',
  POST_REFRESH_TOKEN = 'POST_REFRESH_TOKEN',
  POST_REQUEST_LOGIN = 'POST_REQUEST_LOGIN',
  POST_CONFIRM_LOGIN = 'POST_CONFIRM_LOGIN',
  POST_REQUEST_UPDATE_EMAIL= 'POST_REQUEST_UPDATE_EMAIL',
  POST_CONFIRM_UPDATE_EMAIL= 'POST_CONFIRM_UPDATE_EMAIL'
}